Move terminal contribs into own folder
