+++ src/vs/workbench/contrib/terminal/contrib/accessibility/browser/terminalAccessibilityHelp.ts
