+++ src/vs/workbench/contrib/terminal/contrib/find/browser/terminalFindWidget.ts
