+++ src/vs/workbench/contrib/terminal/contrib/links/browser/terminalLinkHelpers.ts
