+++ src/vs/workbench/contrib/terminal/contrib/links/browser/terminalLinkParsing.ts
