+++ src/vs/workbench/contrib/terminal/test/browser/links/linkTestUtils.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { deepStrictEqual } from 'assert';
-import { ITerminalLinkDetector, TerminalLinkType } from 'vs/workbench/contrib/terminal/browser/links/links';
+import { ITerminalLinkDetector, TerminalLinkType } from 'vs/workbench/contrib/terminal/contrib/links/browser/links';
 import { URI } from 'vs/base/common/uri';
 import { IBufferLine } from 'xterm';
