+++ src/vs/workbench/contrib/terminal/test/browser/links/terminalWordLinkDetector.test.ts
@@ -7,8 +7,8 @@ import { IConfigurationService } from 'vs/platform/configuration/common/configur
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { IProductService } from 'vs/platform/product/common/productService';
-import { ITerminalSimpleLink, TerminalBuiltinLinkType } from 'vs/workbench/contrib/terminal/browser/links/links';
-import { TerminalWordLinkDetector } from 'vs/workbench/contrib/terminal/browser/links/terminalWordLinkDetector';
+import { ITerminalSimpleLink, TerminalBuiltinLinkType } from 'vs/workbench/contrib/terminal/contrib/links/browser/links';
+import { TerminalWordLinkDetector } from 'vs/workbench/contrib/terminal/contrib/links/browser/terminalWordLinkDetector';
 import { assertLinkHelper } from 'vs/workbench/contrib/terminal/test/browser/links/linkTestUtils';
 import { TestProductService } from 'vs/workbench/test/common/workbenchTestServices';
 import { Terminal } from 'xterm';
